requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.3
python-dateutil==2.8.2
pytz==2024.1 